address review and refactor
