+++ src/shared/components/plots/styles.scss
@@ -17,6 +17,10 @@
             width: 240px;
         }
 
+        .cohort-select-div .Select {
+            width: 100%;
+        }
+
         .Select.is-open.is-searchable {
             .Select-value-label {
                 color: #aaaaaa !important;
